remove unused sqlalchemy import; move heat-call pseudo channel into its own module
